dom/worklet/tests/file_dump.html

<!DOCTYPE HTML>
<html>
<head>
  <title>Test for Worklet</title>
  <script type="application/javascript" src="common.js"></script>
</head>
<body>

<script type="application/javascript">

setupTest();

audioWorklet.import("worklet_dump.js")
.then(() => {
  ok(true, "All good!");
  SimpleTest.finish();
});

</script>
</body>
</html>
